requirements.txt
torch==1.9.0
torchvision==0.10.0
opencv-python-headless==4.5.4.58
PyQt5==5.15.4
matplotlib==3.4.3
Pillow==8.3.2
numpy==1.19.5